==> article/04_FirstStage.R <==
################################################################################
#
#                         MCC-CityEurope
#
#                       First-stage models
#
################################################################################

source("00_Packages_Parameters.R")

load("data/Alldata.RData")

#---------------------------
#  Prepare loop
#---------------------------

#----- Prepare parallelisation
ncores <- detectCores()
cl <- makeCluster(max(1, ncores - 2))
registerDoParallel(cl)

#----- Write text file to trace iterations
writeLines(c(""), "temp/logstage1.txt")
cat(as.character(as.POSIXct(Sys.time())), file = "temp/logstage1.txt", 
  append = T)

#----- Variables for age attribution

# Extract and sort age group deaths
deathvars <- sort(grep("deathrate_[[:digit:]]", names(metadata), value = T))

# Extract min, and max of age group
agemin <- substr(deathvars, 11, 12)
agemax <- substr(deathvars, 13, 14)

# Repeat values for each age
ageseqs <- Map(seq, agemin, agemax, by = 1)
agedeaths <- mapply(function(d, a){ 
    do.call(cbind, rep(list(d), length(a)))
  }, metadata[,deathvars], ageseqs)
agedeaths <- do.call(cbind, agedeaths)
ageseq <- unlist(ageseqs, use.names = F)
colnames(agedeaths) <- ageseq
rownames(agedeaths) <- metadata$URAU_CODE

#---------------------------
# Loop on cities
#---------------------------

stage1res <- foreach(dat = iter(dlist), i = iter(seq(dlist)), 
  .packages = c("dlnm", "splines", "MESS")) %dopar% {

  #----- Store iteration (1 every 20)
  if(i %% 20 == 0) cat("\n", "iter = ", i, as.character(Sys.time()), "\n",
    file = "temp/logstage1.txt", append = T)
  
  #----- Prepare model
    
  # Extract ERA5 percentiles
  icode <- metadata[match(cities[i,"city"], metadata$mcc_code), "URAU_CODE"]  
  wholeera5 <- era5series[[icode]]$era5landtmean
    
  # Define crossbasis
  argvar <- list(fun = varfun, degree = vardegree, 
    knots = quantile(wholeera5, varper / 100, na.rm = T),
    Bound = range(wholeera5))
  cb <- crossbasis(dat$era5landtmean, lag = maxlag, argvar = argvar,
    arglag = list(fun = lagfun, knots = lagknots))
  
  # Select outcome: if no age classes select only total age
  #   If no all-cause, select non-external (natural)
  outtype <- ifelse(any(grepl("all", names(dat))), "all", "nonext")
  yvars <- grep(sprintf("%s_", outtype), names(dat), value = T)
  if (length(yvars) == 0) yvars <- outtype
  
  # Determine age ranges
  agerange <- sapply(strsplit(yvars, "_"), "[", 2)
  if (is.na(agerange)){
    amin <- "00"
    amax <- "99"
  } else {
    amin <- substr(agerange, 1, 2)
    amax <- substr(agerange, 3, 4)
  }
  
  # Exclude specific groups
  exclage <- amax < minage
  yvars <- yvars[!exclage]
  amin <- amin[!exclage]
  amax <- amax[!exclage]
  
  # Compute cumulative age groupings
  ytot <- apply(dat[,yvars, drop = F], 2, sum, na.rm = T)
  cumgroups <- cumsumbinning(ytot, mindeath, cutwhenpassed = TRUE)
  
  #----- Loop on each cumulative group
  ageres <- list()
  for (a in unique(cumgroups)){
    
    # Outcome in this cumulative group
    avars <- yvars[cumgroups == a]
    
    # Extract outcome
    y <- rowSums(dat[,avars, drop = F])
    
    # Determine age range
    aamin <- amin[cumgroups == a][1]
    aamax <- tail(amax[cumgroups == a], 1)
    
    ## If aamax < 85 (max age cut point), compute mean age weighted by deathrate
    if (aamax < 85){
      # Select concerned range
      whichgrps <- ageseq >= as.numeric(aamin) & ageseq <= as.numeric(aamax)
      ndeaths <- subset(agedeaths, metadata$mcc_code == cities[i, "city"],
        whichgrps)
      
      # Mean age group weighted by deaths
      meanage <- weighted.mean(ageseq[whichgrps], ndeaths)
    } else { ## Otherwise use life expectancy
      meanage <- metadata[match(cities[i, "city"], metadata$mcc_code), 
        sprintf("lifexp_%s", aamin)] + as.numeric(aamin)
    }
    
    # Run model
    # res <- try(glm(y ~ cb + dow + ns(date, df = 7 * length(unique(year))), 
    #   dat, family = quasipoisson))
    # Exclusion of August 2003
    res <- try(glm(y ~ cb + dow + ns(date, df = 7 * length(unique(year))), 
      dat, family = quasipoisson, subset = !(month == 8 & year == 2003)))
    
    # Some models fail completely (pseudo-weights of IRLS become Inf)
    # Exit current iteration with NAs and a flag for non-convergence
    if (inherits(res, "try-error") || !res$converged){
      nred <- attr(cb, "df")[1]
      ageres[[a]] <- list(coef = rep(NA, nred), 
        vcov = matrix(NA, nred, nred), 
        totdeath = sum(y, na.rm = T), conv = F, ageval = meanage
      )
      names(ageres)[a] <- paste0(aamin,aamax)
      next
    }
    
    # Reduction to overall cumulative
    redall <- crossreduce(cb, res, cen = median(dat$era5landtmean))
    
    # Output
    ageres[[a]] <- list(coef = coef(redall), vcov = vcov(redall), 
      totdeath = sum(y, na.rm = T), conv = res$converged, ageval = meanage
    )
    names(ageres)[a] <- paste0(aamin, aamax)
  }
  
  # Output
  list(tsum = quantile(dat$era5landtmean, predper / 100), modelres = ageres)
}

# Stop parallel
stopCluster(cl)

# Rename
names(stage1res) <- cities$city

# Export
save.image(file = "data/FirstStage.RData")
